requirements.txt
python-dotenv==0.20.0
rich==12.2.0
tweepy==4.8.0
